fix: fall back to CPU if GPU is not available
The current code expected that if the GPU-based tree building is enabled,
that there also is a working GPU available. When it is not, it panics.
This commit adds a fallback to CPU in case selecting a working GPU fails.

It can be verified by running

    FIL_PROOFS_USE_GPU_TREE_BUILDER=1 cargo test --test stacked_vanilla test_stacked_porep_prove_verify

if the machine has no working Nvidia or AMD GPU.

Fixes #1516.